pup: allow configuration via file.
Ready for integration in the main Pedigree tree. This will allow us to remove
all the binaries sitting in the tree at the moment and just use pup in each
checkout. This will reduce the size of the disk image as well as let us test
with a minimal environment.